# [6.0] Cache language files