### Summary of Changes
Add a caching layer to the languageHelper::parseIniFile method.

Benchmark the function on within the system is not easy the the performance gain is in about 1,5 to 20ms depending on language size and server configuration.

My benchmark on this method only for the backend dashboard had the following results.

Without patch: 2,6-3.0ms 
With patch cold cache: 7,6-11ms
With patch warm cache: 0.3-1.11ms

Means a performance gain of 130% to 900% per page load (for language loading).

Using a synthetic 1mb big ini file (35k lines) the following benchmarks apply:
Without patch: 14,6-25,2ms 
With patch cold cache: 32,9-34,5ms
With patch warm cache: 0.5-0,9ms

Means a performance gain of 1500% to 5000% per page load (for language loading).

The Benchmark are taken on a Webserver with production load so might be 
 
The results depend heavily on the server configuration and benefits if opcache is used (which is usually active since a couple of years on most(all) hosting providers).

The cache automatically invalidates it's self as soon as the ini file has a new modification timestamp.

### Testing Instructions
* Apply patch
* Open Frontend and Backend

### Actual result BEFORE applying this Pull Request
* should work

### Expected result AFTER applying this Pull Request
* Should Work
* loads a bit faster.


### Link to documentations
Please select:
- [ ] Documentation link for docs.joomla.org: <link>
- [X] No documentation changes for docs.joomla.org needed

- [ ] Pull Request link for manual.joomla.org: <link>
- [X] No documentation changes for manual.joomla.org needed