@brianteeman 2025-04-06
> The cache automatically invalidates it's self as soon as the ini file has a new modification timestamp.


What about when an override is created?