@brianteeman 2025-04-06
I have tested this item :red_circle: unsuccessfully on 4a7a88cbd97fa5eec7a7cc13e4cd09b05e11285e<br /><br />Applied the PR and no language values are loaded - only constants

Enabled language debug
Warning
: foreach() argument must be of type array|object, string given in
D:\repos\j6\plugins\system\debug\src\DataCollector\LanguageErrorsCollector.php
on line
132<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/45289">issues.joomla.org/tracker/joomla-cms/45289</a>.</sub>