@HLeithner 2025-04-06
funny, because actually the error collector function is broken, can you please.

can you please test 2 things (because I can't replicate it).

first move this line out of the try block, so the exception is not cached and you can see it.
https://github.com/joomla/joomla-cms/blob/103a612249b181ed31886813ec3f26af32b6bd4f/libraries/src/Language/Language.php#L711

second add [] around the text string
https://github.com/joomla/joomla-cms/blob/103a612249b181ed31886813ec3f26af32b6bd4f/libraries/src/Language/Language.php#L719

`$this->errorfiles[$fileName] = ['PHP parser errors :' . $e->getMessage()];`

then your error message shouldn't be warning anymore

third be sure to use the latest version from the "Preserve directory structure" or later commit, since I fixed the that frontend languages are used in backend and vis-a-vis.