@brianteeman 2025-04-06
> first move this line out of the try block, so the exception is not cached and you can see it.

Joomla\Filesystem\Folder::create: Could not create directory. Path: [ROOT]\administrator\cache\language\D:

Look at the paths

![image](https://github.com/user-attachments/assets/ecf11dcd-7078-4d9f-ab61-fd6b021b9158)