@HLeithner 2025-04-06
I have no windows to test, but hope my change fixed it.