@brianteeman 2025-04-07
That change worked.

tested that overrides still work
tested site and admin
tested plugin that has language files in the plugin and not the main languages folder eg plg_sampledate_testing