@brianteeman 2025-04-07
I have tested this item :white_check_mark: successfully on 3acc14c35a3b1cd6e9241aafb55ddb593492b6f3<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/45289">issues.joomla.org/tracker/joomla-cms/45289</a>.</sub>