@HLeithner 2025-04-07
> Reading and parsing a json file is faster.
> 
> like json_decode(file_get_contents('administrator/cache/language/administrator-language-en-GB-lib_joomla.ini.1743321429.json'),true);
> 
> about 30% on my stystem

can you please show me you measurements, because my test says it's about 8-10 times slower using json for empty frontpage and using a 1mb ini files it's 80-100 times slower.