@brbrbr 2025-04-07
Hi,

I found the difference: php is a lot faster with opcache enabled. Without opache json.