@HLeithner 2025-04-07
KISS, if you don't use opcache you don't care about performance so we shouldn't blow up our codebase for nothing.