@exlemor 2025-04-07
I have tested this item :white_check_mark: successfully on 3acc14c35a3b1cd6e9241aafb55ddb593492b6f3<br /><br />I have successfully tested this on 2 sites: non multi-lingual and 1 multi-lingual.


I can honestly say I saw consistently faster speeds however the speed improvement varied from quite a bit from 10% - 26% - I did the battery of test 5 times to get a sense but the large difference I assume comes from 1. the fact that caching is complex and 2. I have fast hosting so the speed difference would be small. Either way I trust you guys! :)<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/45289">issues.joomla.org/tracker/joomla-cms/45289</a>.</sub>