@laoneo 2025-04-25
Just curious, why did you not use the Joomla cache API?