@softforge 2025-04-29
Thank you @HLeithner for the work on this and @brianteeman for testing and helping with the solution and thanks to the other testers who got involved.